code_editor/requirements.txt
flask
flask-sqlalchemy
google-generativeai
python-dotenv